requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai>=1.50.0
requests==2.31.0
python-dotenv>=1.0.0
httpx>=0.27.0
httpx-sse>=0.4.0
mcp>=1.1.0
gunicorn

